Find Closest Point

For a given geometry, find the closest point on that geometry to a point. The closest point may be an intersection, a single point, or unable to be determined.

Usage

closest_point(x, y)

closest_point_haversine(x, y)

Arguments

x

an object of class rsgeo

y

an object of class rs_POINT

Value

An rs_POINT vector

Examples

x <- geom_linestring(1:100, runif(100, 0, 90), rep.int(1:10, 10))
y <- geom_point(runif(10, 0, 90), rnorm(10, 1, 90))
closest_point(x, y)
#> <rs_POINT[10]>
#>  [1] Point(Coord { x: 84.0, y: 11.483372070360929 }))            
#>  [2] Point(Coord { x: 10.0, y: 18.85957977734506 }))             
#>  [3] Point(Coord { x: 77.61921156519773, y: 33.40977802142936 }))
#>  [4] Point(Coord { x: 96.0, y: 75.87950785411522 }))             
#>  [5] Point(Coord { x: 41.0, y: 43.84337919065729 }))             
#>  [6] Point(Coord { x: 58.0, y: 7.718244781717658 }))             
#>  [7] Point(Coord { x: 23.0, y: 11.542766445782036 }))            
#>  [8] Point(Coord { x: 37.0, y: 7.24323813803494 }))              
#>  [9] Point(Coord { x: 55.0, y: 17.83652296056971 }))             
#> [10] Point(Coord { x: 83.51254260478824, y: 52.68513819522518 }))
closest_point_haversine(x, y)
#> <rs_POINT[10]>
#>  [1] Point(Coord { x: 84.0, y: 11.483372070360929 }))             
#>  [2] Point(Coord { x: 10.0, y: 18.85957977734506 }))              
#>  [3] Point(Coord { x: 76.30118630977954, y: 33.758957313218794 }))
#>  [4] Point(Coord { x: 6.0, y: 89.2960596922785 }))                
#>  [5] Point(Coord { x: 41.0, y: 43.84337919065729 }))              
#>  [6] Point(Coord { x: 28.0, y: 2.9271399811841547 }))             
#>  [7] Point(Coord { x: 73.0, y: 80.77474147779867 }))              
#>  [8] Point(Coord { x: 37.0, y: 7.24323813803494 }))               
#>  [9] Point(Coord { x: 55.0, y: 17.83652296056971 }))              
#> [10] Point(Coord { x: 79.33013341522914, y: 52.7517525104448 }))